feat(#16): widen __heap_bump_bytes() from i32 to i64
Every backend's runtime already computed the bump high-water mark in 64
bits — the natives subtract in rax / x0, and wasm's reader had the full
i32 address range. Only the *declared* result was i32, and that was
enough to truncate: a quadratic sweep read 141 MB / 555 MB / -2.09 GB /
202 MB across a 2x-per-step size progression, where only the third of the
four readings looks wrong. The arena is 16 GiB, so the mark passes 2^31 on
exactly the long runs the probe exists to measure — and it is the
sanctioned memory gate (peak RSS varies 12x with transparent hugepages),
so a silently wrapping reading is a silently wrong regression test.

Declared type: checker.FuncSigs and the self-host's asmcore.infer_expr_type
both say i64 now. The self-host previously fell through to `unknown`, which
is what let the two compilers disagree without anything noticing.

Backends: the natives needed nothing — emitHeapBumpBytesRuntime already
returned a full register on both, and the self-host emitters already pushed
a 64-bit slot. wasm needed a real instruction: both operands are i32
linear-memory addresses, so buildHeapBumpBytesBody and its self-host sibling
in wasm_ir.fern subtract in i32 and append i64.extend_i32_u. Zero-, not
sign-extend — a sign-extend would turn back into a negative number the very
reading this widening exists to fix.

Self-host width tracking: irlower's infer_expr_width and lower_i64 learn the
builtin alongside the clock builtins. That is load-bearing rather than
cosmetic — `__heap_bump_bytes() as i32` is the shape every rc probe uses,
and it only routes through lower_i64 + int_wrap when the operand reads 64
there. Without it the whole migrated corpus would mistype on self-host wasm.

Corpus: the ~540 existing probe sites read the mark straight into an i32
exit code or delta, and there is no implicit narrowing, so each call site
gained an explicit `as i32`. That preserves every existing assertion
exactly: all of them measure well under 2 GiB. arm64_asm_bench_run.fern is
the deliberate exception — it prints the mark for a real compile, so it
keeps the full i64 via std/i64 rather than reintroducing the bug in the one
place that currently hits it.

Gates: internal/e2e/rc_heap_bump_i64_test.go. HeapBumpIsI64 binds the probe
to an i64 local and does 64-bit-only arithmetic on all three backends, so a
re-narrowing is a compile error rather than a wrong number;
TestX86_64HeapBumpAbove2GiB bumps the cursor to 2400000032 bytes and reads
it back, which under the old i32 result came out as -1894967264. The
self-host suite gets a matching i64-typed case.

+++ internal/e2e/append_fresh_reclaim_test.go
@@ -75,15 +75,15 @@ function main(): i32 {
     var t: i32 = 0;
     // Each phase: warm up, sample, churn 10x more, sample again. Equal
     // samples == every buffer the 10 extra calls allocated was reclaimed.
-    t = t + churnNone(3); var a1: i32 = __heap_bump_bytes(); t = t + churnNone(10); var a2: i32 = __heap_bump_bytes();
+    t = t + churnNone(3); var a1: i32 = (__heap_bump_bytes() as i32); t = t + churnNone(10); var a2: i32 = (__heap_bump_bytes() as i32);
     if (a2 != a1) { return 11; }
-    t = t + churnOne(3); var b1: i32 = __heap_bump_bytes(); t = t + churnOne(10); var b2: i32 = __heap_bump_bytes();
+    t = t + churnOne(3); var b1: i32 = (__heap_bump_bytes() as i32); t = t + churnOne(10); var b2: i32 = (__heap_bump_bytes() as i32);
     if (b2 != b1) { return 12; }
-    t = t + churnTwo(3); var c1: i32 = __heap_bump_bytes(); t = t + churnTwo(10); var c2: i32 = __heap_bump_bytes();
+    t = t + churnTwo(3); var c1: i32 = (__heap_bump_bytes() as i32); t = t + churnTwo(10); var c2: i32 = (__heap_bump_bytes() as i32);
     if (c2 != c1) { return 13; }
-    t = t + churnLoop(3); var d1: i32 = __heap_bump_bytes(); t = t + churnLoop(10); var d2: i32 = __heap_bump_bytes();
+    t = t + churnLoop(3); var d1: i32 = (__heap_bump_bytes() as i32); t = t + churnLoop(10); var d2: i32 = (__heap_bump_bytes() as i32);
     if (d2 != d1) { return 14; }
-    t = t + churnDistinct(3); var e1: i32 = __heap_bump_bytes(); t = t + churnDistinct(10); var e2: i32 = __heap_bump_bytes();
+    t = t + churnDistinct(3); var e1: i32 = (__heap_bump_bytes() as i32); t = t + churnDistinct(10); var e2: i32 = (__heap_bump_bytes() as i32);
     if (e2 != e1) { return 15; }
 `
